Add `pytest-cov` and new inv task
This raises a question about how we run tests and what we should consider when
checking coverage.

Currently this coverage check does _not_ get run in CI. We have to run it
manually. This is because we only have integration/regression tests and we don't
have any infra in place to automatically run those as on NSIDC hardware (w/
access to data, ancillary fields, etc) yet.

Ideally as much code as possible is unit tested (not requiring access to
non-repo data) and the test coverage should reflect that. However, there is
still value in regression/integration tests that require access to non-repo
data.

Maybe: ignore regression/integration tests and run the coverage check w/ just
unit test? Need to think some more on this.