Merge Scott's code that fixes the doy invalid ice mask generation
The original ice mask we pull from `seaice_ecdr` is actually a bitmask that
includes to daily and monthly icemask for each doy. This updated code extracts
the correct daily mask and also unifies the landmask with our existing
ancillary files for `seaice_ecdr`, which are different from the `seaice_cdr` landmask.