Make error handling around missing data issues more generic
For 0080 and 0001 specifically. These datasets produce netcdf files for each day
that have groups within for each platform (e.g. a group called `F17` for the tbs
from F17). Sometimes there is no data for the platform we want data for, but the
file still exists. This causes an OSError and was not properly handled by our code.