Remove unused/OBE kwargs related to ancillary sources
`ancillary_source` was a very common kwarg that was originally intended to be
used to test differences between CDR v4 and 5. The code was never fully
implemented, and did not work as intended.

For CDR v6, there will be updated and new ancillary data.

This major version change will remove the concept of ancillary sources, removing
a broken feature and streamlining the code in the process.

If we need to add support for different ancillary sources in the future, we
should promote it to a top-level configuration that does not need to be passed
through so many pieces of code, and ensure it is well tested!